Remove <Esc><Esc> mapping
This mapping can be inserted character A, B, C or D unexpectedly.

Cursor key sends <Esc>OA (or <Esc>OB, <Esc>OC, <Esc>OD) in xterm.
If any cursor key hits immediately after the <Esc>, it is construed as <Esc><Esc>OA.

Unfortunately, an mapping for <Esc><Esc> is applied and 'OA' is left.
'OA' is a command to insert a new line with character A above the cursor.