Replace up/down-line-or-search by up/down-line-or-beginning-search
up-line-or-search/down-line-or-search will get history by the first word in the buffer.
By this design, it prevents to search history with command and arguments.

To search a history with arguments, it needs to use up/down-line-or-beginning-search.
It seems that up/down-line-or-beginning-search works as same as history-search-end.

This commit nearly reverts 8f4893eeb2a97939805ed02a1c8e8d7bbf0ba7ce.