TERM should not be set manually
In most cases, TERM is set by automatically. (e.g. "linux" in Linux console)
We should respect the default value.

In particular, it can be changed by a configuration when using a terminal emulator.
TERM should be set in this manner instead of a shell configuration.

Only the terminal knows its own capability. The shell does not.

Let us illustrate this issue using a simple example.
TERM is overwritten by the shell to "xterm-256color" every time logging-in with SSH.
It is done even though the local terminal is a Linux console.
As a result, the terminal capability between local and remote is mismatched.