Keymap without Ctrl key should not be repeatable
Single key keymap tends to collision with application keymap.

For example, 'j' has a meaning of 'Move to the next window' in tmux and 'Scroll forward next line' in Vim.
In that case, 'jj' will be inputted if a user want to move to next window and scroll down the application with in the window.

However actual result is moving to the next next window if 'j' key is repeatable in tmux.

In contrast with this, a keymap with Ctrl key have less incidence of collision with application keymap.
It can prevent false operation with repeatability of keymap.