Avoid downloading same contents as much as possible
Containers with same ID and different tags exist in some cases.
These containers are listed in the neighborhood at output of 'docker images' on its default sort order.
It is very likely to cause downloading same content because 'docker pull' is run in sequential and parallrel.
To avoid this issue, shuffle the output of 'docker images'.